Psql: disk-eater

+++ server/getData.sh
@@ -16,14 +16,17 @@ name="$2"
 }
 
 zipShp "http://opendata.iprpraha.cz/CUR/OVZ/OVZ_Klima_Osluneni_p/WGS_84/OVZ_Klima_Osluneni_p_shp.zip" OVZ_Klima_Osluneni_p
-#zipShp "http://opendata.iprpraha.cz/CUR/HM/HM_Ekola_den_p/WGS_84/HM_Ekola_den_p_shp.zip" HM_Ekola_den_p
-#zipShp "http://opendata.iprpraha.cz/CUR/URK/URK_SS_VyuzitiZakl_p/WGS_84/URK_SS_VyuzitiZakl_p_shp.zip" URK_SS_VyuzitiZakl_p
+zipShp "http://opendata.iprpraha.cz/CUR/HM/HM_Ekola_den_p/WGS_84/HM_Ekola_den_p_shp.zip" HM_Ekola_den_p
+zipShp "http://opendata.iprpraha.cz/CUR/URK/URK_SS_VyuzitiZakl_p/WGS_84/URK_SS_VyuzitiZakl_p_shp.zip" URK_SS_VyuzitiZakl_p
 
 
-psql -c "drop database knp_praha"
-psql -c "create database knp_praha"
-psql -d knp_praha -c "CREATE EXTENSION postgis;"
-psql -d knp_praha -c "GRANT ALL PRIVILEGES ON DATABASE knp_praha TO knp"
+if [ 0"$1" -eq 1 ]
+ then
+ psql -c "drop database knp_praha"
+ psql -c "create database knp_praha"
+ psql -d knp_praha -c "CREATE EXTENSION postgis;"
+ psql -d knp_praha -c "GRANT ALL PRIVILEGES ON DATABASE knp_praha TO knp"
+fi
 
 
 for f in  data/src/* #data/*.json
@@ -32,14 +35,27 @@ for f in  data/src/* #data/*.json
   then #not a directory
   continue
  fi
+#temporary
+#if [ ! "$f" = "data/src/HM_Ekola_den_p" ]
+# then
+# continue
+#fi
+
  echo $f;
  tablename=`echo $f|sed -e 's/^data\/src\///g'`
- #echo $tablename
- #ogr2ogr -f "PostgreSQL" PG:"dbname=knp_praha user=postgres" data/OVZ_Klima_Osluneni_p.json -nln $tablename -append
- #psql -d knp_praha -c "SELECT ST_AsText(ST_GeomFromGeoJSON()))";
- shp2pgsql "$f"/*.shp "$tablename" | psql -d knp_praha
+ shp2pgsql "$f"/*.shp "$tablename"_source | psql -d knp_praha
 
- psql -d knp_praha -c "CREATE INDEX $tablename_gix ON $tablename USING GIST (geom);"
+
+psql -d knp_praha -c "CREATE TABLE $tablename AS SELECT  *, (ST_DUMP(geom)).geom::geometry(Polygon,0) AS new_geom FROM ""$tablename""_source"
+
+ psql -d knp_praha -c "ALTER TABLE $tablename DROP COLUMN geom;"
+ psql -d knp_praha -c "ALTER TABLE $tablename RENAME COLUMN new_geom TO geom;"
+ psql -d knp_praha -c "VACUUM ANALYZE $tablename;"
+
+
+ psql -d knp_praha -c "CREATE INDEX ""$tablename""_gix ON $tablename USING GIST (geom);"
+
+ psql -d knp_praha -c "DROP TABLE ""$tablename""_source;"
 
  done
